pyproject.toml
[project]
name = "realtime-audio-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sounddevice>=0.5.0",
    "numpy>=2.1.2",
    "websockets>=14.1",
    "soundfile>=0.13.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.0.1",
    "scipy>=1.15.0",
    "soundcard>=0.4.3",
]
